requirements.txt
streamlit
openai
llama-index
chromadb
pandas
python-dotenv
sentence-transformers
